Updating the document as per the review comments.
You must provide a description of how you will implement the following
(mandatory) stage 2 requirements:

App integrates two or more Google services. Google service integrations
can be a part of Google Play Services or Firebase.
App provides a widget to provide relevant information to the user on
the home screen.
Must implement at least one of the three
If it regularly pulls or sends data to/from a web service or API, app
updates data in its cache at regular intervals using a SyncAdapter.
OR
If it needs to pull or send data to/from a web service or API only
once, or on a per request basis (such as a search application), app
uses an IntentService to do so.
OR
It it performs short duration, on-demand requests (such as search), app
uses an AsyncTask.
  RESUBMIT PROJECT